pyproject.toml
[project]
name = "ai-blog-outline-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
]
